pyproject.toml
[project]
name = "mcp-script-runner"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastmcp",
  "python-dotenv>=1.1.0",
]
